+++ serde-jackson/src/test/groovy/io/micronaut/serde/jackson/annotation/SerdeJsonFormatSpec.groovy
@@ -1,22 +1,12 @@
 package io.micronaut.serde.jackson.annotation
 
-
-import io.micronaut.serde.jackson.JsonCompileSpec
-import io.micronaut.serde.config.SerdeConfiguration.NumericTimeUnit
+import io.micronaut.serde.jackson.JsonFormatSpec
 import spock.lang.Unroll
 
-import java.sql.Timestamp
 import java.time.Instant
-import java.time.LocalDate
 import java.time.LocalDateTime
-import java.time.LocalTime
-import java.time.OffsetDateTime
-import java.time.Year
-import java.time.ZoneId
-import java.time.ZoneOffset
-import java.time.ZonedDateTime
 
-class JsonFormatSpec extends JsonCompileSpec {
+class SerdeJsonFormatSpec extends JsonFormatSpec {
 
     void "test disable validation"() {
         when:
@@ -193,11 +183,10 @@ class Test {
         BigDecimal | new BigDecimal("100000.12") | [pattern: '$###,###.###']                  | '{"value":"$100,000.12"}'
         BigInteger | new BigInteger("100000")    | [pattern: '$###,###.###']                  | '{"value":"$100,000"}'
         BigInteger | new BigInteger("100000")    | [pattern: '$###,###.###']                  | '{"value":"$100,000"}'
-
     }
 
     @Unroll
-    void "test json format for date #type and settings #settings"() {
+    void "INSTANT + SQL DATE test json format for date #type and settings #settings"() {
         given:
         def context = buildContext('test.Test', """
 package test;
@@ -232,62 +221,7 @@ class Test {
         where:
         type           | value                                     | settings                                | resolver
         Instant        | Instant.now()                             | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ"] | { Instant i -> i.toEpochMilli() }
-        Date           | new Date()                                | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ"] | { Date d -> d.time }
         java.sql.Date  | new java.sql.Date(2021, 9, 15)            | [pattern: "yyyy-MM-dd"]                 | { java.sql.Date d -> d }
-        Timestamp      | new Timestamp(System.currentTimeMillis()) | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ"] | { Timestamp d -> d }
-        LocalTime      | LocalTime.now()                           | [pattern: "HH:mm:ss"]                   | { LocalTime i -> i.toSecondOfDay() }
-        LocalDate      | LocalDate.now()                           | [pattern: "yyyy-MM-dd"]                 | { LocalDate d -> d }
-        LocalDateTime  | LocalDateTime.now()                       | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSS"]  | { LocalDateTime i -> i.toInstant(ZoneOffset.from(ZoneOffset.UTC)).toEpochMilli() }
-        ZonedDateTime  | ZonedDateTime.now()                       | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ"] | { ZonedDateTime i -> i.toInstant().toEpochMilli() }
-        OffsetDateTime | OffsetDateTime.now()                      | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ"] | { OffsetDateTime i -> i.toInstant().toEpochMilli() }
-        Year           | Year.of(2021)                             | [pattern: "yyyy"]                       | { Year y -> y }
     }
 
-    @Unroll
-    void "test deserialize json number format for date #type"() {
-        given:
-        def context = buildContext('test.Test', """
-package test;
-
-import io.micronaut.serde.annotation.Serdeable;
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.OptBoolean;
-
-@Serdeable
-class Test {
-    @JsonFormat(${settings.collect { "$it.key=\"$it.value\"" }.join(",")})
-    private $type.name value;
-    public void setValue($type.name value) {
-        this.value = value;
-    }
-    public $type.name getValue() {
-        return value;
-    }
-}
-""", [:], ['micronaut.serde.numeric-time-unit': timeUnit])
-
-        def jsonString = """
-{
-    "value": ${value}
-}
-"""
-        def read = jsonMapper.readValue(jsonString, typeUnderTest)
-
-        expect:
-        resolver(read.value) == expected
-
-        cleanup:
-        context.close()
-
-        where:
-        type           | timeUnit                     | value                         | settings                                                 | resolver                                                                 | expected
-        Instant        | NumericTimeUnit.SECONDS      | "1640995200"                  | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ", timezone: "UTC"] | { Instant i -> i.getEpochSecond() }                                      | 1640995200
-        Date           | NumericTimeUnit.MILLISECONDS | "1640995200000"               | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ", timezone: "UTC"] | { Date d -> d.getTime() }                                                | 1640995200000
-        Timestamp      | NumericTimeUnit.MILLISECONDS | "1640995200000"               | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ", timezone: "UTC"] | { Timestamp t -> t.getTime()}                                            | 1640995200000
-        LocalDate      | NumericTimeUnit.SECONDS      | "19974"                       | [pattern: "yyyy-MM-dd", timezone: "UTC"]                 | { LocalDate d -> d.toString() }                                          | "2024-09-08"
-        LocalDateTime  | NumericTimeUnit.SECONDS      | "\"2024-10-18T23:06:24.722\"" | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSS", timezone: "UTC"]  | { LocalDateTime t -> t.atZone(ZoneId.of("UTC")).toInstant().toString() } | "2024-10-18T23:06:24.722Z"
-        ZonedDateTime  | NumericTimeUnit.SECONDS      | "1640995200"                  | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ", timezone: "UTC"] | { ZonedDateTime t -> t.toString() }                                      | "2022-01-01T00:00Z"
-        OffsetDateTime | NumericTimeUnit.SECONDS      | "1640995200"                  | [pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ", timezone: "UTC"] | { OffsetDateTime t -> t.toString() }                                     | "2022-01-01T00:00Z"
-        Year           | NumericTimeUnit.SECONDS      | "2024"                        | [pattern: "yyyy", timezone: "UTC"]                       | { Year y -> y.toString() }                                               | "2024"
-    }
 }
